move sidebar types to /types
