+++ frontend/src/types/sidebar.ts
